requirements.txt
Flask>=2.0.0
Werkzeug>=2.0.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
PyMySQL==1.0.2
marshmallow==3.13.0
Flask-JWT-Extended==4.3.1
python-dotenv==0.19.0
Flask-SocketIO==5.1.1
flasgger==0.9.5
flask-cors==4.0.1
setuptools>=42.0.0

